move InMemoryRepo to data
